+++ apps/ui-dev/component/navigations/sidebar/NestedView.jsx
@@ -1,0 +1,46 @@
+import SettingsAccessibilityFilled from "@elf-framework/icon/SettingsAccessibilityFilled";
+import { Sidebar, VBox } from "@elf-framework/ui";
+
+export function NestedView() {
+  return (
+    <VBox
+      style={{
+        gap: 30,
+        backgroundColor: "var(--color-gray-2)",
+        padding: "30px 150px",
+        fontSize: 13,
+      }}
+    >
+      <div
+        style={{
+          display: "flex",
+          gap: 30,
+          flexDirection: "column",
+          padding: [20, 40],
+          // justifyContent: "center",
+          alignItems: "center",
+        }}
+      >
+        <Sidebar
+          items={[
+            "Section title",
+            "-",
+            {
+              title: "xxx",
+              link: "xxx",
+              icon: <SettingsAccessibilityFilled />,
+            },
+            { title: "xxx", link: "xxx", selected: true },
+            "-",
+            "Hello World",
+            { title: "blue" },
+            {
+              title: "group",
+              items: [{ title: "xxxx" }, { title: "xxxx" }, { title: "xxxx" }],
+            },
+          ]}
+        />
+      </div>
+    </VBox>
+  );
+}